sec_keypad.idc: remove builtIn
While review 51328 properly fixed the keyboard cursor orientation, it
also introduced an alt key regression by utilizing the buggy, and
superfluous, builtIn = 1 directive.

With builtIn = 1, ALT_RIGHT and ALT_LEFT are folded into one event,
which makes it possible for applications, such as terminal apps, to
distinguish between ALT as a "fn" key and ALT as in a TTY ALT key. With
builtIn = 0, the desire behavior is restored, and the fix of review
51328 continues to work.

Change-Id: I8f058fdce03efe653eeaa3c8b35ef55b649029a8